Optimise tests
