+++ test/e2e/hazelcast_cache_persistence_test.go
@@ -2,18 +2,15 @@ package e2e
 
 import (
 	"context"
-	"fmt"
 	"strconv"
 	. "time"
 
-	hz "github.com/hazelcast/hazelcast-go-client"
 	. "github.com/onsi/ginkgo/v2"
 	. "github.com/onsi/gomega"
 	v1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/types"
 
 	hazelcastv1alpha1 "github.com/hazelcast/hazelcast-platform-operator/api/v1alpha1"
-	"github.com/hazelcast/hazelcast-platform-operator/internal/protocol/codec"
 	hazelcastconfig "github.com/hazelcast/hazelcast-platform-operator/test/e2e/config/hazelcast"
 )
 
@@ -133,44 +130,3 @@ var _ = Describe("Hazelcast Cache Config with Persistence", Label("cache_persist
 	})
 
 })
-
-func validateCacheEntriesPortForward(h *hazelcastv1alpha1.Hazelcast, localPort, cacheName string, entryCount int) {
-	stopChan := portForwardPod(h.Name+"-0", h.Namespace, localPort+":5701")
-	defer closeChannel(stopChan)
-	cl := newHazelcastClientPortForward(context.Background(), h, localPort)
-	cli := hz.NewClientInternal(cl)
-	for i := 0; i < entryCount; i++ {
-		key, err := cli.EncodeData(fmt.Sprintf("mykey%d", i))
-		Expect(err).To(BeNil())
-		value := fmt.Sprintf("myvalue%d", i)
-		getRequest := codec.EncodeCacheGetRequest("/hz/"+cacheName, key, nil)
-		resp, err := cli.InvokeOnKey(context.Background(), getRequest, key, nil)
-		pairs := codec.DecodeCacheGetResponse(resp)
-		Expect(err).To(BeNil())
-		data, err := cli.DecodeData(pairs)
-		Expect(err).To(BeNil())
-		Expect(fmt.Sprintf("%v", data)).Should(Equal(value))
-	}
-}
-
-func fillCachePortForward(h *hazelcastv1alpha1.Hazelcast, cacheName, localPort string, entryCount int) {
-	stopChan := portForwardPod(h.Name+"-0", h.Namespace, localPort+":5701")
-	defer closeChannel(stopChan)
-	cl := newHazelcastClientPortForward(context.Background(), h, localPort)
-	cli := hz.NewClientInternal(cl)
-
-	for _, mi := range cli.OrderedMembers() {
-		configRequest := codec.EncodeCacheGetConfigRequest("/hz/"+cacheName, cacheName)
-		_, _ = cli.InvokeOnMember(context.Background(), configRequest, mi.UUID, nil)
-	}
-
-	for i := 0; i < entryCount; i++ {
-		key, err := cli.EncodeData(fmt.Sprintf("mykey%d", i))
-		Expect(err).To(BeNil())
-		value, err := cli.EncodeData(fmt.Sprintf("myvalue%d", i))
-		Expect(err).To(BeNil())
-		cpr := codec.EncodeCachePutRequest("/hz/"+cacheName, key, value, nil, false, 0)
-		_, err = cli.InvokeOnKey(context.Background(), cpr, key, nil)
-		Expect(err).To(BeNil())
-	}
-}
