+++ test/e2e/hazelcast_map_test.go
@@ -40,16 +40,6 @@ var _ = Describe("Hazelcast Map Config", Label("map"), func() {
 		GinkgoWriter.Printf("Aftereach end time is %v\n", Now().String())
 	})
 
-	It("should create Map Config", Label("fast"), func() {
-		setLabelAndCRName("hm-1")
-		hazelcast := hazelcastconfig.Default(hzLookupKey, ee, labels)
-		CreateHazelcastCR(hazelcast)
-
-		m := hazelcastconfig.DefaultMap(mapLookupKey, hazelcast.Name, labels)
-		Expect(k8sClient.Create(context.Background(), m)).Should(Succeed())
-		assertMapStatus(m, hazelcastcomv1alpha1.MapSuccess)
-	})
-
 	It("should create Map Config with correct default values", Label("fast"), func() {
 		setLabelAndCRName("hm-2")
 		hazelcast := hazelcastconfig.Default(hzLookupKey, ee, labels)
