+++ test/e2e/hazelcast_replicatedmap_test.go
@@ -29,16 +29,6 @@ var _ = Describe("Hazelcast ReplicatedMap Config", Label("replicatedmap"), func(
 		GinkgoWriter.Printf("Aftereach end time is %v\n", Now().String())
 	})
 
-	It("should create ReplicatedMap Config", Label("fast"), func() {
-		setLabelAndCRName("hrm-1")
-		hazelcast := hazelcastconfig.Default(hzLookupKey, ee, labels)
-		CreateHazelcastCR(hazelcast)
-
-		rm := hazelcastconfig.DefaultReplicatedMap(rmLookupKey, hazelcast.Name, labels)
-		Expect(k8sClient.Create(context.Background(), rm)).Should(Succeed())
-		assertDataStructureStatus(rmLookupKey, hazelcastcomv1alpha1.DataStructureSuccess, &hazelcastcomv1alpha1.ReplicatedMap{})
-	})
-
 	It("should create ReplicatedMap Config with correct default values", Label("fast"), func() {
 		setLabelAndCRName("hrm-2")
 		hazelcast := hazelcastconfig.Default(hzLookupKey, ee, labels)
