+++ test/e2e/hazelcast_topic_test.go
@@ -27,16 +27,6 @@ var _ = Describe("Hazelcast Topic Config", Label("topic"), func() {
 		GinkgoWriter.Printf("Aftereach end time is %v\n", Now().String())
 	})
 
-	It("should create Topic Config", Label("fast"), func() {
-		setLabelAndCRName("ht-1")
-		hazelcast := hazelcastconfig.Default(hzLookupKey, ee, labels)
-		CreateHazelcastCR(hazelcast)
-
-		topic := hazelcastconfig.DefaultTopic(topicLookupKey, hazelcast.Name, labels)
-		Expect(k8sClient.Create(context.Background(), topic)).Should(Succeed())
-		assertDataStructureStatus(topicLookupKey, hazelcastcomv1alpha1.DataStructureSuccess, &hazelcastcomv1alpha1.Topic{})
-	})
-
 	It("should create Topic Config with correct default values", Label("fast"), func() {
 		setLabelAndCRName("ht-2")
 		hazelcast := hazelcastconfig.Default(hzLookupKey, ee, labels)
